Added animations
We created the animation class and added a few methods too the assets
class and player class that allow us to create and run animations.
We changed mario to link and built the main animations with a for loop and
the idle animation with a delay on frame one to delay the blink animation.